Revert "#114 design : Add login context & Social Login Screen"
This reverts commit 5dfffef9c8acaf16381ab8277386da4ae5d79b2e.
