+++ App.tsx
@@ -20,7 +20,6 @@ import { GestureHandlerRootView } from "react-native-gesture-handler";
 import HomeScreen from "./src/pages/Home";
 import { Coord } from "react-native-nmap";
 import SplashScreen from "react-native-splash-screen";
-import { LoginProvider } from "./src/common/Context";
 
 export type AppProps = {
   Home: any;
@@ -38,18 +37,16 @@ const App = (): JSX.Element => {
 
   return (
     <GestureHandlerRootView style={{ flex: 1 }}>
-      <LoginProvider>
-        <NavigationContainer>
-          <Stack.Navigator
-            screenOptions={() => ({
-              headerShown: false,
-            })}
-          >
-            <Stack.Screen name="Home" component={HomeScreens} />
-            <Stack.Screen name="Login" component={LoginScreen} />
-          </Stack.Navigator>
-        </NavigationContainer>
-      </LoginProvider>
+      <NavigationContainer>
+        <Stack.Navigator
+          screenOptions={() => ({
+            headerShown: false,
+          })}
+        >
+          <Stack.Screen name="Home" component={HomeScreens} />
+          <Stack.Screen name="Login" component={LoginScreen} />
+        </Stack.Navigator>
+      </NavigationContainer>
     </GestureHandlerRootView>
   );
 };
