--- src/common/Context.tsx
@@ -1,24 +1,0 @@
-import React, { ReactNode, createContext, useEffect, useState } from 'react'
-import { getAccessToken } from './storage';
-
-export const LoginContext = createContext({
-  isLogin: false,
-  setLogin: (value: boolean) => { }
-})
-
-export const LoginProvider = ({ children }: { children: ReactNode }) => {
-  const [isLogin, setIsLogin] = useState<boolean>(false);
-  const checkLogin = async () => {
-    const token = await getAccessToken();
-    if (token) setIsLogin(true);
-  }
-  const setLogin = (value: boolean) => {
-    setIsLogin(value);
-  }
-  useEffect(() => { checkLogin();},[])
-  return (
-    <LoginContext.Provider value={{ isLogin, setLogin }}>
-      {children}
-    </LoginContext.Provider>
-  )
-}
